refactor: text 대신 expired 매개변수 사용하도록 변경
