--- app/src/main/java/woowacourse/payments/ui/component/ExpiredInput.kt
@@ -1,63 +1,0 @@
-package woowacourse.payments.ui.component
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.tooling.preview.Preview
-import woowacourse.payments.R
-import woowacourse.payments.domain.Expired
-import woowacourse.payments.ui.theme.AndroidpaymentsTheme
-
-@Composable
-fun ExpiredInput(
-    expired: Expired?,
-    onExpiredChange: (Expired?) -> Unit,
-    modifier: Modifier = Modifier,
-    showValidationError: Boolean = false,
-) {
-    var text by remember { mutableStateOf("") }
-
-    Column(modifier = modifier) {
-        OutlinedTextField(
-            value = text,
-            onValueChange = { newText ->
-                val filteredText = newText.filter { it.isDigit() }.take(4)
-                text = filteredText
-                onExpiredChange(Expired(filteredText))
-            },
-            visualTransformation = ExpiredVisualTransformation(groupSize = 2, delimiter = " / "),
-            modifier = Modifier.fillMaxWidth(0.5f),
-            label = { Text(text = stringResource(R.string.expired_label)) },
-            placeholder = {
-                Text(
-                    text = stringResource(R.string.expired_placeholder),
-                    color = Color.LightGray,
-                )
-            },
-            isError = showValidationError && (expired?.isValid != true),
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-        )
-    }
-}
-
-@Composable
-@Preview(showBackground = true)
-fun ExpiredInputPreview() {
-    AndroidpaymentsTheme {
-        ExpiredInput(
-            expired = null,
-            onExpiredChange = { },
-        )
-    }
-}
